rename run module to start and update README with CLI and SDK documentation
